\name{SparseData-package}
\alias{SparseData-package}
\alias{SparseData}
\docType{package}
\title{
SparseData: Efficient Manipulation of Sparse Data
}
\description{
Extends the expression set paradigm for sparse matrices, offering differential testing and correlation/distance measures.
}
\details{
\tabular{ll}{
Package: \tab SparseData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-09\cr
License: \tab GPL (>= 3)\cr
Depends: \tab methods\cr
}

Most important functions: \code{\link{newSparseDataSet}},
    \code{\link{calculateMeans}}, \code{\link{calculateTStats}},
    \code{\link{sparseCov}} 

}
\author{
Michael Love <love@molgen.mpg.de>
}
\keyword{ package }

